+++ crates/completion/src/render.rs
@@ -194,10 +194,7 @@ impl<'a> Render<'a> {
                     local_name,
                 )
                 .kind(CompletionItemKind::UnresolvedReference)
-                .add_import(
-                    import_to_add,
-                    self.ctx.completion.config.should_resolve_additional_edits_immediately(),
-                )
+                .add_import(import_to_add, self.ctx.completion.config.resolve_edits_immediately())
                 .build();
                 return Some(item);
             }
@@ -252,10 +249,7 @@ impl<'a> Render<'a> {
 
         let item = item
             .kind(kind)
-            .add_import(
-                import_to_add,
-                self.ctx.completion.config.should_resolve_additional_edits_immediately(),
-            )
+            .add_import(import_to_add, self.ctx.completion.config.resolve_edits_immediately())
             .set_documentation(docs)
             .set_ref_match(ref_match)
             .build();
