+++ crates/completion/src/render/function.rs
@@ -47,10 +47,7 @@ impl<'a> FunctionRender<'a> {
             .set_deprecated(self.ctx.is_deprecated(self.func))
             .detail(self.detail())
             .add_call_parens(self.ctx.completion, self.name, params)
-            .add_import(
-                import_to_add,
-                self.ctx.completion.config.should_resolve_additional_edits_immediately(),
-            )
+            .add_import(import_to_add, self.ctx.completion.config.resolve_edits_immediately())
             .build()
     }
